+++ esp32c3-embassy/src/random.rs
@@ -8,49 +8,56 @@
 
 //! Random numbers generator
 
-use rand_core::CryptoRng;
-use rand_core::RngCore;
+use core::convert::Infallible;
 
-use esp_hal::rng::Rng;
+use rand_core::Rng as _;
+use rand_core::TryCryptoRng;
+use rand_core::TryRng;
 
-/// A wrapper for ESP random number generator that implement traits form
+use esp_hal::rng::Rng as EspRng;
+
+/// A wrapper for ESP random number generator that implements traits form
 /// `rand_core`
 #[derive(Clone)]
-pub struct RngWrapper(Rng);
+pub struct RngWrapper(EspRng);
 
-impl From<Rng> for RngWrapper {
-    fn from(rng: Rng) -> Self {
+impl From<EspRng> for RngWrapper {
+    fn from(rng: EspRng) -> Self {
         Self(rng)
     }
 }
 
-impl RngCore for RngWrapper {
-    fn next_u32(&mut self) -> u32 {
-        self.0.random()
+impl TryRng for RngWrapper {
+    type Error = Infallible;
+
+    fn try_next_u32(&mut self) -> Result<u32, Self::Error> {
+        Ok(self.0.random())
     }
 
-    fn next_u64(&mut self) -> u64 {
-        u32_pair_to_u64(self.next_u32(), self.next_u32())
+    fn try_next_u64(&mut self) -> Result<u64, Self::Error> {
+        Ok(u32_pair_to_u64(self.next_u32(), self.next_u32()))
     }
 
-    fn fill_bytes(&mut self, dest: &mut [u8]) {
+    fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), Self::Error> {
         for value in dest.iter_mut() {
-            let [random_value, _, _, _] = self.next_u32().to_ne_bytes();
+            let [random_value, _, _, _] = self.next_u32().to_be_bytes();
             *value = random_value;
         }
+
+        Ok(())
     }
 }
 
-impl CryptoRng for RngWrapper {}
+impl TryCryptoRng for RngWrapper {}
 
 /// Join a pair of `u32` into a `u64`
-#[allow(
-    clippy::many_single_char_names,
-    clippy::min_ident_chars,
-    reason = "This is still readable"
-)]
 fn u32_pair_to_u64(first: u32, second: u32) -> u64 {
-    let [a, b, c, d] = first.to_ne_bytes();
-    let [e, f, g, h] = second.to_ne_bytes();
-    u64::from_ne_bytes([a, b, c, d, e, f, g, h])
+    #![expect(
+        clippy::many_single_char_names,
+        clippy::min_ident_chars,
+        reason = "This is still readable"
+    )]
+    let [a, b, c, d] = first.to_be_bytes();
+    let [e, f, g, h] = second.to_be_bytes();
+    u64::from_be_bytes([a, b, c, d, e, f, g, h])
 }